Crate rusoto_connect

The Amazon Connect API Reference provides descriptions, syntax, and usage examples for each of the Amazon Connect actions, data types, parameters, and errors. Amazon Connect is a cloud-based contact center solution that makes it easy to set up and manage a customer contact center and provide reliable customer engagement at any scale.

Throttling limits for the Amazon Connect API operations:

For the GetMetricData and GetCurrentMetricData operations, a RateLimit of 5 per second, and a BurstLimit of 8 per second.

For all other operations, a RateLimit of 2 per second, and a BurstLimit of 5 per second.

You can request an increase to the throttling limits by submitting a Amazon Connect service limits increase form. You must be signed in to your AWS account to access the form.

If you're using the service, you're probably looking for ConnectClient and Connect.

Structs

ConnectClient
A client for the Amazon Connect API.

CreateUserRequest

CreateUserResponse

Credentials
The credentials to use for federation.

CurrentMetric
A CurrentMetric object that contains the Name and Unit for the metric.

CurrentMetricData
A CurrentMetricData object.

CurrentMetricResult
A CurrentMetricResult object.

DeleteUserRequest

DescribeUserHierarchyGroupRequest

DescribeUserHierarchyGroupResponse

DescribeUserHierarchyStructureRequest

DescribeUserHierarchyStructureResponse

DescribeUserRequest

DescribeUserResponse

Dimensions
A Dimensions object that includes the Channel and Queue for the metric.

Filters
The filter, either channel or queues, to apply to the metric results retrieved.

GetContactAttributesRequest

GetContactAttributesResponse

GetCurrentMetricDataRequest

GetCurrentMetricDataResponse

GetFederationTokenRequest

GetFederationTokenResponse

GetMetricDataRequest

GetMetricDataResponse

HierarchyGroup
A HierarchyGroup object that contains information about a hierarchy group in your Amazon Connect instance.

HierarchyGroupSummary
A HierarchyGroupSummary object that contains information about the hierarchy group, including ARN, Id, and Name.

HierarchyLevel
A HierarchyLevel object that contains information about the levels in a hierarchy group, including ARN, Id, and Name.

HierarchyPath
A HierarchyPath object that contains information about the levels of the hierarchy group.

HierarchyStructure
A HierarchyStructure object that contains information about the hierarchy group structure.

HistoricalMetric
A HistoricalMetric object that contains the Name, Unit, Statistic, and Threshold for the metric.

HistoricalMetricData
A HistoricalMetricData object than contains a Metric and a Value.

HistoricalMetricResult
The metrics data returned from a GetMetricData operation.

ListRoutingProfilesRequest

ListRoutingProfilesResponse

ListSecurityProfilesRequest

ListSecurityProfilesResponse

ListUserHierarchyGroupsRequest

ListUserHierarchyGroupsResponse

ListUsersRequest

ListUsersResponse

QueueReference
A QueueReference object that contains the the QueueId and ARN for the queue resource for which metrics are returned.

RoutingProfileSummary
A RoutingProfileSummary object that contains information about a routing profile, including ARN, Id, and Name.

SecurityProfileSummary
A SecurityProfileSummary object that contains information about a security profile, including ARN, Id, Name.

StartOutboundVoiceContactRequest

StartOutboundVoiceContactResponse

StopContactRequest

StopContactResponse

Threshold
A Threshold object that includes a comparison and ThresholdValue to compare to. Used with service level metrics.

UpdateContactAttributesRequest

UpdateContactAttributesResponse

UpdateUserHierarchyRequest

UpdateUserIdentityInfoRequest

UpdateUserPhoneConfigRequest

UpdateUserRoutingProfileRequest

UpdateUserSecurityProfilesRequest

User
A User object that contains information about a user account in your Amazon Connect instance, including configuration settings.

UserIdentityInfo
A UserIdentityInfo object that contains information about the user's identity, including email address, first name, and last name.

UserPhoneConfig
A UserPhoneConfig object that contains information about the user phone configuration settings.

UserSummary
A UserSummary object that contains Information about a user, including ARN, Id, and user name.

Enums

CreateUserError
Errors returned by CreateUser

DeleteUserError
Errors returned by DeleteUser

DescribeUserError
Errors returned by DescribeUser

DescribeUserHierarchyGroupError
Errors returned by DescribeUserHierarchyGroup

DescribeUserHierarchyStructureError
Errors returned by DescribeUserHierarchyStructure

GetContactAttributesError
Errors returned by GetContactAttributes

GetCurrentMetricDataError
Errors returned by GetCurrentMetricData

GetFederationTokenError
Errors returned by GetFederationToken

GetMetricDataError
Errors returned by GetMetricData

ListRoutingProfilesError
Errors returned by ListRoutingProfiles

ListSecurityProfilesError
Errors returned by ListSecurityProfiles

ListUserHierarchyGroupsError
Errors returned by ListUserHierarchyGroups

ListUsersError
Errors returned by ListUsers

StartOutboundVoiceContactError
Errors returned by StartOutboundVoiceContact

StopContactError
Errors returned by StopContact

UpdateContactAttributesError
Errors returned by UpdateContactAttributes

UpdateUserHierarchyError
Errors returned by UpdateUserHierarchy

UpdateUserIdentityInfoError
Errors returned by UpdateUserIdentityInfo

UpdateUserPhoneConfigError
Errors returned by UpdateUserPhoneConfig

UpdateUserRoutingProfileError
Errors returned by UpdateUserRoutingProfile

UpdateUserSecurityProfilesError
Errors returned by UpdateUserSecurityProfiles

Traits

Connect
Trait representing the capabilities of the Amazon Connect API. Amazon Connect clients implement this trait.