Crate rusoto_connect

The Amazon Connect API Reference provides descriptions, syntax, and usage examples for each of the Amazon Connect actions, data types, parameters, and errors. Amazon Connect is a cloud-based contact center solution that makes it easy to set up and manage a customer contact center and provide reliable customer engagement at any scale.

If you're using the service, you're probably looking for ConnectClient and Connect.

Structs

ConnectClient
A client for the Amazon Connect API.

CreateUserRequest

CreateUserResponse

Credentials
The credentials to use for federation.

DeleteUserRequest

DescribeUserHierarchyGroupRequest

DescribeUserHierarchyGroupResponse

DescribeUserHierarchyStructureRequest

DescribeUserHierarchyStructureResponse

DescribeUserRequest

DescribeUserResponse

GetFederationTokenRequest

GetFederationTokenResponse

HierarchyGroup
A HierarchyGroup object that contains information about a hierarchy group in your Amazon Connect instance.

HierarchyGroupSummary
A HierarchyGroupSummary object that contains information about the hierarchy group, including ARN, Id, and Name.

HierarchyLevel
A HierarchyLevel object that contains information about the levels in a hierarchy group, including ARN, Id, and Name.

HierarchyPath
A HierarchyPath object that contains information about the levels of the hierarchy group.

HierarchyStructure
A HierarchyStructure object that contains information about the hierarchy group structure.

ListRoutingProfilesRequest

ListRoutingProfilesResponse

ListSecurityProfilesRequest

ListSecurityProfilesResponse

ListUserHierarchyGroupsRequest

ListUserHierarchyGroupsResponse

ListUsersRequest

ListUsersResponse

RoutingProfileSummary
A RoutingProfileSummary object that contains information about a routing profile, including ARN, Id, and Name.

SecurityProfileSummary
A SecurityProfileSummary object that contains information about a security profile, including ARN, Id, Name.

StartOutboundVoiceContactRequest

StartOutboundVoiceContactResponse

StopContactRequest

StopContactResponse

UpdateUserHierarchyRequest

UpdateUserIdentityInfoRequest

UpdateUserPhoneConfigRequest

UpdateUserRoutingProfileRequest

UpdateUserSecurityProfilesRequest

User
A User object that contains information about a user account in your Amazon Connect instance, including configuration settings.

UserIdentityInfo
A UserIdentityInfo object that contains information about the user's identity, including email address, first name, and last name.

UserPhoneConfig
A UserPhoneConfig object that contains information about the user phone configuration settings.

UserSummary
A UserSummary object that contains Information about a user, including ARN, Id, and user name.

Enums

CreateUserError
Errors returned by CreateUser

DeleteUserError
Errors returned by DeleteUser

DescribeUserError
Errors returned by DescribeUser

DescribeUserHierarchyGroupError
Errors returned by DescribeUserHierarchyGroup

DescribeUserHierarchyStructureError
Errors returned by DescribeUserHierarchyStructure

GetFederationTokenError
Errors returned by GetFederationToken

ListRoutingProfilesError
Errors returned by ListRoutingProfiles

ListSecurityProfilesError
Errors returned by ListSecurityProfiles

ListUserHierarchyGroupsError
Errors returned by ListUserHierarchyGroups

ListUsersError
Errors returned by ListUsers

StartOutboundVoiceContactError
Errors returned by StartOutboundVoiceContact

StopContactError
Errors returned by StopContact

UpdateUserHierarchyError
Errors returned by UpdateUserHierarchy

UpdateUserIdentityInfoError
Errors returned by UpdateUserIdentityInfo

UpdateUserPhoneConfigError
Errors returned by UpdateUserPhoneConfig

UpdateUserRoutingProfileError
Errors returned by UpdateUserRoutingProfile

UpdateUserSecurityProfilesError
Errors returned by UpdateUserSecurityProfiles

Traits

Connect
Trait representing the capabilities of the Amazon Connect API. Amazon Connect clients implement this trait.