Struct rusoto_connect::ConnectClient

pub struct ConnectClient { /* fields omitted */ }

A client for the Amazon Connect API.

Methods

impl ConnectClient

pub fn new(region: Region) -> ConnectClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ConnectClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,