Trait rusoto_connect::Connect

pub trait Connect {
    fn create_user(
        &self, 
        input: CreateUserRequest
    ) -> RusotoFuture<CreateUserResponse, CreateUserError>;

    fn delete_user(
        &self, 
        input: DeleteUserRequest
    ) -> RusotoFuture<(), DeleteUserError>;

    fn describe_user(
        &self, 
        input: DescribeUserRequest
    ) -> RusotoFuture<DescribeUserResponse, DescribeUserError>;

    fn describe_user_hierarchy_group(
        &self, 
        input: DescribeUserHierarchyGroupRequest
    ) -> RusotoFuture<DescribeUserHierarchyGroupResponse, DescribeUserHierarchyGroupError>;

    fn describe_user_hierarchy_structure(
        &self, 
        input: DescribeUserHierarchyStructureRequest
    ) -> RusotoFuture<DescribeUserHierarchyStructureResponse, DescribeUserHierarchyStructureError>;

    fn get_federation_token(
        &self, 
        input: GetFederationTokenRequest
    ) -> RusotoFuture<GetFederationTokenResponse, GetFederationTokenError>;

    fn list_routing_profiles(
        &self, 
        input: ListRoutingProfilesRequest
    ) -> RusotoFuture<ListRoutingProfilesResponse, ListRoutingProfilesError>;

    fn list_security_profiles(
        &self, 
        input: ListSecurityProfilesRequest
    ) -> RusotoFuture<ListSecurityProfilesResponse, ListSecurityProfilesError>;

    fn list_user_hierarchy_groups(
        &self, 
        input: ListUserHierarchyGroupsRequest
    ) -> RusotoFuture<ListUserHierarchyGroupsResponse, ListUserHierarchyGroupsError>;

    fn list_users(
        &self, 
        input: ListUsersRequest
    ) -> RusotoFuture<ListUsersResponse, ListUsersError>;

    fn start_outbound_voice_contact(
        &self, 
        input: StartOutboundVoiceContactRequest
    ) -> RusotoFuture<StartOutboundVoiceContactResponse, StartOutboundVoiceContactError>;

    fn stop_contact(
        &self, 
        input: StopContactRequest
    ) -> RusotoFuture<StopContactResponse, StopContactError>;

    fn update_user_hierarchy(
        &self, 
        input: UpdateUserHierarchyRequest
    ) -> RusotoFuture<(), UpdateUserHierarchyError>;

    fn update_user_identity_info(
        &self, 
        input: UpdateUserIdentityInfoRequest
    ) -> RusotoFuture<(), UpdateUserIdentityInfoError>;

    fn update_user_phone_config(
        &self, 
        input: UpdateUserPhoneConfigRequest
    ) -> RusotoFuture<(), UpdateUserPhoneConfigError>;

    fn update_user_routing_profile(
        &self, 
        input: UpdateUserRoutingProfileRequest
    ) -> RusotoFuture<(), UpdateUserRoutingProfileError>;

    fn update_user_security_profiles(
        &self, 
        input: UpdateUserSecurityProfilesRequest
    ) -> RusotoFuture<(), UpdateUserSecurityProfilesError>;
}

Trait representing the capabilities of the Amazon Connect API. Amazon Connect clients implement this trait.

Required Methods

fn create_user(
&self,
input: CreateUserRequest
) -> RusotoFuture<CreateUserResponse, CreateUserError>

Creates a new user account in your Amazon Connect instance.

fn delete_user(
&self,
input: DeleteUserRequest
) -> RusotoFuture<(), DeleteUserError>

Deletes a user account from Amazon Connect.

fn describe_user(
&self,
input: DescribeUserRequest
) -> RusotoFuture<DescribeUserResponse, DescribeUserError>

Returns a User object that contains information about the user account specified by the UserId.

fn describe_user_hierarchy_group(
&self,
input: DescribeUserHierarchyGroupRequest
) -> RusotoFuture<DescribeUserHierarchyGroupResponse, DescribeUserHierarchyGroupError>

Returns a HierarchyGroup object that includes information about a hierarchy group in your instance.

fn describe_user_hierarchy_structure(
&self,
input: DescribeUserHierarchyStructureRequest
) -> RusotoFuture<DescribeUserHierarchyStructureResponse, DescribeUserHierarchyStructureError>

Returns a HiearchyGroupStructure object, which contains data about the levels in the agent hierarchy.

fn get_federation_token(
&self,
input: GetFederationTokenRequest
) -> RusotoFuture<GetFederationTokenResponse, GetFederationTokenError>

Retrieves a token for federation.

fn list_routing_profiles(
&self,
input: ListRoutingProfilesRequest
) -> RusotoFuture<ListRoutingProfilesResponse, ListRoutingProfilesError>

Returns an array of RoutingProfileSummary objects that includes information about the routing profiles in your instance.

fn list_security_profiles(
&self,
input: ListSecurityProfilesRequest
) -> RusotoFuture<ListSecurityProfilesResponse, ListSecurityProfilesError>

Returns an array of SecurityProfileSummary objects that contain information about the security profiles in your instance, including the ARN, Id, and Name of the security profile.

fn list_user_hierarchy_groups(
&self,
input: ListUserHierarchyGroupsRequest
) -> RusotoFuture<ListUserHierarchyGroupsResponse, ListUserHierarchyGroupsError>

Returns a UserHierarchyGroupSummaryList, which is an array of HierarchyGroupSummary objects that contain information about the hierarchy groups in your instance.

fn list_users(
&self,
input: ListUsersRequest
) -> RusotoFuture<ListUsersResponse, ListUsersError>

Returns a UserSummaryList, which is an array of UserSummary objects.

fn start_outbound_voice_contact(
&self,
input: StartOutboundVoiceContactRequest
) -> RusotoFuture<StartOutboundVoiceContactResponse, StartOutboundVoiceContactError>

The StartOutboundVoiceContact operation initiates a contact flow to place an outbound call to a customer.

There is a throttling limit placed on usage of the API that includes a RateLimit of 2 per second, and a BurstLimit of 5 per second.

If you are using an IAM account, it must have permission to the connect:StartOutboundVoiceContact action.

fn stop_contact(
&self,
input: StopContactRequest
) -> RusotoFuture<StopContactResponse, StopContactError>

Ends the contact initiated by the StartOutboundVoiceContact operation.

If you are using an IAM account, it must have permission to the connect:StopContact action.

fn update_user_hierarchy(
&self,
input: UpdateUserHierarchyRequest
) -> RusotoFuture<(), UpdateUserHierarchyError>

Assigns the specified hierarchy group to the user.

fn update_user_identity_info(
&self,
input: UpdateUserIdentityInfoRequest
) -> RusotoFuture<(), UpdateUserIdentityInfoError>

Updates the identity information for the specified user in a UserIdentityInfo object, including email, first name, and last name.

fn update_user_phone_config(
&self,
input: UpdateUserPhoneConfigRequest
) -> RusotoFuture<(), UpdateUserPhoneConfigError>

Updates the phone configuration settings in the UserPhoneConfig object for the specified user.

fn update_user_routing_profile(
&self,
input: UpdateUserRoutingProfileRequest
) -> RusotoFuture<(), UpdateUserRoutingProfileError>

Assigns the specified routing profile to a user.

fn update_user_security_profiles(
&self,
input: UpdateUserSecurityProfilesRequest
) -> RusotoFuture<(), UpdateUserSecurityProfilesError>

Update the security profiles assigned to the user.